tests: Tweak touch test

+++ tests/cases/060_volume/010_touch/test.ps1
@@ -1,4 +1,4 @@
-# SUMMARY: Simple test that we can create (touch) a file on the host
+# SUMMARY: Create (touch) a file on volume mount and check on the host
 # LABELS:
 # REPEAT:
 
@@ -6,16 +6,18 @@ Set-PSDebug -Trace 2
 
 $fileName = "foobar"
 
-# Make sure the 
-Remove-Item -Path $fileName -Force
+if (Test-Path $fileName) {
+    Remove-Item -Path $fileName -Force
+}
 
 $p = [string]$pwd.Path
-docker run --platform linux --rm -v  $p`:/test alpine:3.7 touch /test/$fileName
+docker run --platform linux --rm -v  $p`:/test alpine:3.7 sh -c "touch /test/$fileName"
 if ($lastexitcode -ne 0) { 
     exit 1
 }
 
-if (Test-Path $fileName) {
+if (Test-Path $fileName -PathType leaf) {
+    Remove-Item -Path $fileName -Force
     exit 0
 }
 exit 1
